cleanup: zero clippy warnings (lib + bins + tests), eval byte-identical
Resolved every clippy warning across the crate (0 on `cargo clippy --all-targets`) while keeping the
frozen champion's eval byte-for-byte identical (tests/eval + tests/nnue_accumulator green throughout):

- machine-applicable fixes via `clippy --fix` (map_or, matches!, redundant casts, ? operator,
  to_vec/first/is_multiple_of, needless mut/parens, collapsible if, redundant bindings, etc.).
- Default for MoveList; TT replacement boolean simplified to the proven-equivalent
  `depth >= cur.depth || (!same && gen != generation)`; matches! conversions in syzygy/book;
  perft type alias; analyze forced-move if-let (drops an unwrap-after-is_some + a stale dev comment).
- crate-level `#![allow(clippy::needless_range_loop)]` (lib.rs) with rationale: the NNUE/eval/SEE
  numeric hot paths use explicit index loops on purpose — reordering a float dot-product would change
  the f32 result and break net parity. Documented intent, not an oversight.
- uci main(): `#[allow(unused_assignments)]` for the intentional ponder-abort searcher restore
  (a macro can't be conditional per call site; the restore is correct at sites that reuse it).

Verified: cargo clippy --all-targets = 0 warnings; cargo test = 148 passed; eval unchanged.

+++ src/search.rs
@@ -358,7 +358,7 @@ impl Searcher {
         // 2. Syzygy Tablebase Root Probe
         if opts.syzygy {
             if let Some(tb) = &self.tb {
-                if pos.castling == 0 && pos.all.count_ones() as u32 <= tb.max_pieces() {
+                if pos.castling == 0 && pos.all.count_ones() <= tb.max_pieces() {
                     if let Some((mv, wdl)) = tb.probe_root(pos) {
                         let score: i32 = match wdl {
                             pyrrhic_rs::WdlProbeResult::Win => 900_000,
